Moved all utils in own files in `utils` folder like reacts port. Keeps folder structure less noisy
