+++ src/lib/hooks/useHover/index.svelte.ts
@@ -1,14 +1,12 @@
 import type { OpenChangeReason } from '$lib/types.js';
-import {
-	contains,
-	createAttribute,
-	getDocument,
-	isMouseLikePointerType,
-	noop,
-} from '$lib/utils.js';
 import { isElement } from '@floating-ui/utils/dom';
 import type { FloatingContext } from '../useFloating/index.svelte.js';
 import type { ElementProps } from '../useInteractions/index.svelte.js';
+import { createAttribute } from '$lib/utils/create-attribute.js';
+import { isMouseLikePointerType } from '$lib/utils/is-mouse-like-pointer-type.js';
+import { noop } from '$lib/utils/noop.js';
+import { getDocument } from '$lib/utils/get-document.js';
+import { contains } from '$lib/utils/contains.js';
 
 interface DelayOptions {
 	/**
