+++ src/lib/utils/create-attribute.ts
@@ -1,0 +1,6 @@
+/**
+ * Returns a data attribute with the given name.
+ */
+export function createAttribute(name: string) {
+	return `data-floating-ui-${name}`;
+}
