+++ src/lib/utils/noop.ts
@@ -1,0 +1,5 @@
+/**
+ * A function that does nothing.
+ * Useful as a default value for optional callbacks.
+ */
+export function noop() {}
