+++ src/lib/utils/style-object-to-string.ts
@@ -1,0 +1,8 @@
+/**
+ * Converts a JavaScript object representing CSS styles into a string.
+ */
+export function styleObjectToString(styleObject: Partial<ElementCSSInlineStyle['style']>) {
+	return Object.entries(styleObject)
+		.map(([key, value]) => `${key}: ${value};`)
+		.join(' ');
+}
